Rewrite class component as functional component and clear the code
I migrate to functional component because it is more clear to work in.
I also clear the App to work with modular functions: where the parts of some feature work independently.

This big change in code cleanup helped with the simple example: Before, when I was trying to create an input validator, I mixed a function that was tightly coupled into a function that didn't need the resources that function performed. See function: #checkInputTypoBySection (old)

However, now I just need to call the #isItemFailed function to get whether the item failed the test or not, regardless of where it was used.

If I want to validate multiple items, I simply call #getItemsWithValidations, which returns an array of mapped items with their fault values already inserted.

This modularization of the components helps me to reuse it later in any location, the only requirement being that it be the same structure that the application expects.